# Increase encoding::decode_varint performance