I managed to increase the performance of encoding::decode_varint by removing some branching and getting rid of the slow decode loop.

This was mostly done though trial-and-error, but I suspect it has something to do with a branching/inline interaction. The result, at least on my machine, is sizable performance gains for small varints (I imagine the most common case, at least as far as tags go) and diminishing performance gains as the varints get larger.

Note: My machine is a old AMD FX-8150 running Linux WSL2. I would be interested to see how the performance is on a fancy new cloud machine.

```
varint/small/encode     time:   [256.46 ns 257.98 ns 259.63 ns]
                        change: [-7.8342% -1.1468% +5.8832%] (p = 0.79 > 0.05)
                        No change in performance detected.

varint/small/decode     time:   [178.23 ns 179.15 ns 180.15 ns]
                        change: [-53.589% -50.993% -48.859%] (p = 0.00 < 0.05)
                        Performance has improved.

varint/small/encoded_len
                        time:   [204.08 ns 208.19 ns 214.88 ns]
                        change: [-4.9152% +1.3815% +7.2996%] (p = 0.69 > 0.05)
                        No change in performance detected.

varint/medium/encode    time:   [1.3080 us 1.3151 us 1.3242 us]
                        change: [-2.3332% +0.0991% +2.1227%] (p = 0.94 > 0.05)
                        No change in performance detected.

varint/medium/decode    time:   [524.90 ns 530.74 ns 539.69 ns]
                        change: [-35.461% -32.583% -28.303%] (p = 0.00 < 0.05)
                        Performance has improved.

varint/medium/encoded_len
                        time:   [203.92 ns 205.00 ns 206.15 ns]
                        change: [-7.3905% +0.0838% +9.0508%] (p = 0.99 > 0.05)
                        No change in performance detected.

varint/large/encode     time:   [2.5958 us 2.6075 us 2.6210 us]
                        change: [-9.8420% -3.6040% +0.6323%] (p = 0.27 > 0.05)
                        No change in performance detected.

varint/large/decode     time:   [869.92 ns 877.55 ns 886.62 ns]
                        change: [-18.593% -14.844% -9.7530%] (p = 0.00 < 0.05)
                        Performance has improved.

varint/large/encoded_len
                        time:   [203.44 ns 204.58 ns 205.87 ns]
                        change: [-3.4912% +1.4936% +7.5105%] (p = 0.68 > 0.05)
                        No change in performance detected.

varint/mixed/encode     time:   [1.7580 us 1.7728 us 1.7959 us]
                        change: [-1.6477% +2.2962% +7.9238%] (p = 0.53 > 0.05)
                        No change in performance detected.

varint/mixed/decode     time:   [1.1937 us 1.1986 us 1.2038 us]
                        change: [-21.372% -18.918% -16.197%] (p = 0.00 < 0.05)
                        Performance has improved.

varint/mixed/encoded_len
                        time:   [203.10 ns 205.56 ns 209.07 ns]
                        change: [-3.4880% -0.5121% +2.5100%] (p = 0.77 > 0.05)
                        No change in performance detected.
```

For reference the "before" decode benchmarks:

```
varint/small/decode     time:   [357.20 ns 359.74 ns 363.04 ns]
varint/medium/decode    time:   [800.92 ns 805.41 ns 810.49 ns]
varint/large/decode     time:   [1.0376 us 1.0480 us 1.0636 us]
varint/mixed/decode     time:   [1.4746 us 1.4838 us 1.4948 us]
```

Taking a page from @quininer (#519), I also tried out `buf.get_u8()`. Unfortunately, its still quite a bit slower than `bytes.get_unchecked(index)` + `buf.advance(len)`.